remove server.go
