+++ config/config.go
@@ -77,3 +77,12 @@ func validateDatabaseConfig(cfg *db.Config) error {
 	}
 	return nil
 }
+
+// ServerConfig defines the configuration for the HTTP server
+type ServerConfig struct {
+	Port               int    `mapstructure:"port"`
+	ServerMode         string `mapstructure:"server-mode"`
+	SessionMaxAge      int    `mapstructure:"session-max-age"`
+	RefreshTokenMaxAge int    `mapstructure:"refresh-token-max-age"`
+	ServiceName        string `mapstructure:"service-name"`
+}
